requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.0
python-socketio[client]==5.4.0
RPi.GPIO==0.7.0